refactor pkg structure
