+++ cmd/arena/commands/get.go
@@ -20,10 +20,10 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"github.com/kubeflow/arena/util"
+	"github.com/kubeflow/arena/pkg/util"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/kubeflow/arena/util/helm"
+	"github.com/kubeflow/arena/pkg/util/helm"
 	"github.com/spf13/cobra"
 
 	"io"
