+++ cmd/arena/commands/gpu_info_test.go
@@ -2,9 +2,10 @@ package commands
 
 import (
 	"fmt"
-	"github.com/kubeflow/arena/util"
 	"strings"
 	"testing"
+
+	"github.com/kubeflow/arena/pkg/util"
 )
 
 func TestQueryMetricByPrometheus(t *testing.T) {
