+++ cmd/arena/commands/logviewer.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/kubeflow/arena/util/helm"
+	"github.com/kubeflow/arena/pkg/util/helm"
 	"github.com/spf13/cobra"
 )
