+++ cmd/arena/commands/serve_delete.go
@@ -17,7 +17,7 @@ package commands
 import (
 	"os"
 
-	"github.com/kubeflow/arena/util/helm"
+	"github.com/kubeflow/arena/pkg/util/helm"
 	"github.com/spf13/cobra"
 )
