+++ cmd/arena/commands/submit_horovod.go
@@ -19,8 +19,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/kubeflow/arena/util"
-	"github.com/kubeflow/arena/util/helm"
+	"github.com/kubeflow/arena/pkg/util"
+	"github.com/kubeflow/arena/pkg/util/helm"
 	"github.com/spf13/cobra"
 )
