+++ cmd/arena/commands/tensorboard.go
@@ -19,7 +19,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/kubeflow/arena/util"
+	"github.com/kubeflow/arena/pkg/util"
 	log "github.com/sirupsen/logrus"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
